Refactor admin table code
