+++ components/table/admin-table/AdminTable.tsx
@@ -1,39 +1,18 @@
 import React, { useState, useEffect, useRef } from 'react';
-import { Button, Input, Space, Tag, Table, TableProps, InputRef } from 'antd';
+import { Button, Input, Space, TableProps, InputRef } from 'antd';
 import type {
   ColumnType,
   FilterValue,
   SorterResult,
 } from 'antd/es/table/interface';
-import {
-  PlusOutlined,
-  SearchOutlined,
-  MinusOutlined,
-  ArrowRightOutlined,
-  UserOutlined,
-} from '@ant-design/icons';
-import * as XLSX from 'xlsx';
-import { saveAs } from 'file-saver';
+import { SearchOutlined } from '@ant-design/icons';
 import useSWR from 'swr';
 import Highlighter from 'react-highlight-words';
 import { QueriedAdminData } from 'bookem-shared/src/types/database';
-import { SearchContainter, TableContainer } from '@/styles/table.styles';
-import Link from 'next/link';
-import { ObjectId } from 'mongodb';
-import CreateEventPopupWindow from '@/components/Forms/CreateEventPopupWindow';
+import { convertAdminDataToRowData } from '@/utils/table-utils';
 import type { FilterConfirmProps } from 'antd/es/table/interface';
-
-interface AdminRowData {
-  key: number;
-  firstName: string;
-  lastName: string;
-  email: string;
-  phone: string;
-  role: string;
-  id: ObjectId;
-}
-
-type DataIndex = keyof AdminRowData;
+import { AdminDataIndex, AdminRowData } from '@/utils/table-types';
+import AdminTableImpl from './AdminTableImpl';
 
 const fetcher = (url: string) => fetch(url).then(res => res.json());
 
@@ -66,7 +45,7 @@ const AdminTable = () => {
   const handleSearch = (
     selectedKeys: string[],
     confirm: (param?: FilterConfirmProps) => void,
-    dataIndex: DataIndex
+    dataIndex: AdminDataIndex
   ) => {
     confirm();
     setSearchText(selectedKeys[0]);
@@ -80,7 +59,7 @@ const AdminTable = () => {
    */
   // Function to get column search properties
   const getColumnSearchProps = (
-    dataIndex: DataIndex
+    dataIndex: AdminDataIndex
   ): ColumnType<AdminRowData> => ({
     // Filter dropdown configuration
     filterDropdown: ({
@@ -195,7 +174,7 @@ const AdminTable = () => {
   }, [data, isLoading]);
 
   // check for errors and loading
-  if (error) return <div>Failed to load event table</div>;
+  if (error) return <div>Failed to load admin table</div>;
   if (isLoading) return <div>Loading...</div>;
 
   // function that determines what the table looks like after a search by the user
@@ -215,157 +194,16 @@ const AdminTable = () => {
     setIsFilter(true);
   };
 
-  // function to export what is on the table at the time to an excel file
-  const handleExport = () => {
-    const workbook = XLSX.utils.table_to_book(
-      document.querySelector('#table-container')
-    );
-    const excelBuffer = XLSX.write(workbook, {
-      bookType: 'xlsx',
-      type: 'array',
-    });
-    const blob = new Blob([excelBuffer], {
-      type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-    });
-    saveAs(blob, 'events.xlsx');
-  };
-
-  // Define columns for the Ant Design table
-  const columns: any = [
-    {
-      title: 'First Name',
-      dataIndex: 'firstName',
-      key: 'firstName',
-      ...getColumnSearchProps('firstName'),
-    },
-    {
-      title: 'Last Name',
-      dataIndex: 'lastName',
-      key: 'lastName',
-      ...getColumnSearchProps('lastName'),
-    },
-    {
-      title: 'Email',
-      dataIndex: 'email',
-      key: 'email',
-      ...getColumnSearchProps('email'),
-    },
-    {
-      title: 'Phone',
-      dataIndex: 'phone',
-      key: 'phone',
-      ...getColumnSearchProps('phone'),
-    },
-    { title: 'Role', dataIndex: 'role', key: 'role' },
-    {
-      title: 'View',
-      dataIndex: '',
-      key: 'x',
-      render: () => (
-        <a>
-          <Button
-            icon={<ArrowRightOutlined rev={undefined} />}
-            onClick={() => {}}
-            style={{
-              width: 30,
-              borderRadius: 70,
-              marginLeft: 30,
-              backgroundColor: 'white',
-              color: 'black',
-              borderColor: 'black',
-            }}></Button>
-        </a>
-      ),
-    },
-  ];
-
   return (
     <>
-      {showPopup && <CreateEventPopupWindow setShowPopup={setShowPopup} />}
-
-      <SearchContainter style={{ marginLeft: 0 }}>
-        <Input
-          ref={searchInput}
-          placeholder={'Search'}
-          onChange={e => (e.target.value ? [e.target.value] : [])}
-          prefix={<SearchOutlined rev={undefined} />}
-          style={{ width: 1000, display: 'flex' }}
-        />
-        <Button
-          icon={<PlusOutlined rev={undefined} />}
-          onClick={() => {}}
-          style={{
-            width: 50,
-            marginLeft: 10,
-            borderRadius: 100,
-            backgroundColor: 'darkgray',
-            color: 'whitesmoke',
-          }}></Button>
-        <Button
-          icon={<MinusOutlined rev={undefined} />}
-          onClick={() => {}}
-          style={{
-            width: 50,
-            marginLeft: 10,
-            borderRadius: 100,
-            backgroundColor: 'darkgray',
-            color: 'whitesmoke',
-          }}></Button>
-        <Button
-          icon={<UserOutlined rev={undefined} />}
-          onClick={() => {}}
-          style={{
-            width: 250,
-            marginLeft: 10,
-            backgroundColor: 'darkgray',
-            color: 'whitesmoke',
-          }}>
-          Your Account
-        </Button>
-      </SearchContainter>
-
-      <TableContainer>
-        <div id="table-container">
-          <Table
-            dataSource={isFiltering ? filterTable : dataForTable}
-            onChange={handleChange}
-            columns={columns}
-            pagination={false}
-            scroll={{ y: 550 }}
-          />
-        </div>
-      </TableContainer>
+      <AdminTableImpl
+        getColumnSearchProps={getColumnSearchProps}
+        sortedInfo={sortedInfo}
+        handleChange={handleChange}
+        dataForTable={dataForTable}
+        searchInput={searchInput}
+      />
     </>
   );
 };
-
-// const convertEventDataToRowData = (data: QueriedVolunteerEventDTO[]) => {
-//   const result = data.map((event, index) => {
-//     return {
-//       key: index,
-//       firstName: 'Manish',
-//       lastName: 'Acharya',
-//       email: 'manish@gmail.com',
-//       phone: '1234567890',
-//       role: 'Super Admin',
-//     };
-//   });
-//   return result;
-// };
-
-const convertAdminDataToRowData = (data: QueriedAdminData[]) => {
-  const result = data.map((user, index) => {
-    return {
-      key: index,
-      firstName: user.firstName,
-      lastName: user.lastName,
-      email: user.email,
-      phone: user.phone,
-      role: user.status,
-      id: user._id,
-    };
-  });
-  return result;
-};
-
 export default AdminTable;
