+++ pages/settings.tsx
@@ -3,7 +3,7 @@ import { signOut } from 'next-auth/react';
 import styled from 'styled-components';
 import { PageLayout, PageTitle } from '@/styles/table.styles';
 import { Admin } from 'mongodb';
-import AdminTable from '@/components/table/AdminTable';
+import AdminTable from '@/components/table/admin-table/AdminTable';
 import { Button } from 'antd';
 import { LogoutOutlined } from '@ant-design/icons';
