+++ utils/table-utils.ts
@@ -1,8 +1,9 @@
 import {
   QueriedVolunteerEventDTO,
   QueriedVolunteerProgramData,
+  QueriedAdminData,
 } from 'bookem-shared/src/types/database';
-import { EventRowData, ProgramRowData } from './table-types';
+import { EventRowData, ProgramRowData, AdminRowData } from './table-types';
 
 export const convertEventDataToRowData = (
   data: QueriedVolunteerEventDTO[]
@@ -28,6 +29,21 @@ export const convertProgramDataToRowData = (
       key: program._id.toString(),
       numEvents: program.events.length,
       description: program.description ?? '',
+    };
+  });
+  return result;
+};
+
+export const convertAdminDataToRowData = (data: QueriedAdminData[]) => {
+  const result = data.map((user, index) => {
+    return {
+      key: index,
+      firstName: user.firstName,
+      lastName: user.lastName,
+      email: user.email,
+      phone: user.phone,
+      role: user.status,
+      id: user._id,
     };
   });
   return result;
